loan_management/requirements.txt
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==4.1.0
bleach==6.0.0
blinker==1.5
cachetools==5.3.0
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.25.2
colorama==0.4.6
comm==0.1.2
confluent-kafka==2.2.0
contourpy==1.0.7
cryptography==39.0.1
cvlib==0.2.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.4
djangorestframework==3.14.0
dlib==19.24.0
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.1.0
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
flatbuffers==1.12
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
grpcio==1.34.1
gspread==5.10.0
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.22.0
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
imutils==0.5.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.0.3
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
kiwisolver==1.4.4
kombu==5.3.2
lazy_loader==0.2
libclang==16.0.0
lightgbm==4.0.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mongoengine==0.27.0
multidict==6.0.4
mysqlclient==2.1.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
progressbar==2.5
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
PyJWT==2.8.0
pylibsrtp==0.8.0
pymongo==4.4.1
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.6
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==4.6.0
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
semver==2.13.0
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.4
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.18.1
termcolor==1.1.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
vine==5.0.0
virtualenv==20.19.0
waitress==2.1.2
watchdog==2.2.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.12.1
xgboost==1.7.6
yarl==1.9.2
zipp==3.13.0
